+++ src/lib-imap/imap-envelope.h
@@ -5,7 +5,7 @@ typedef struct _MessagePartEnvelopeData MessagePartEnvelopeData;
 
 void imap_envelope_parse_header(Pool pool, MessagePartEnvelopeData **data,
 				const char *name,
-				const char *value, unsigned int value_len);
+				const char *value, size_t value_len);
 
 void imap_envelope_write_part_data(MessagePartEnvelopeData *data,
 				   TempString *str);
